requirements.txt
langchain
langchain_community
langchain-classic
langchain-google-genai
python-dotenv